+++ pkg/wasi/hook/link.dart
@@ -6,8 +6,6 @@ import 'package:wasm_tools/hooks.dart';
 
 void main(List<String> args) => link(args, (input, output) async {
   if (input.config.buildWasmComponent) {
-    output.dependencies.add(Uri.parse('hook/wasm_abi.json'));
-
     final uses = input.recordedUses!;
     final usedWorlds = <String>{};
 
@@ -27,6 +25,8 @@ void main(List<String> args) => link(args, (input, output) async {
     }
 
     for (final used in usedWorlds) {
+      output.dependencies.add(.parse(used));
+
       output.assets.webAssemblyComponents.add(
         WasmComponentAsset(
           encoded: json.decode(