+++ pkg/wasm_tools/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1-preview.0
 
 - Improve witgen.
+- Export types when needed.
 
 ## 0.1.0-preview.0